requirements.txt
pypdf
pypdf[image]
Pillow
argparse
tqdm
PyMuPDF